круд и список обращений готов + пагинация

+++ app/app/Http/Controllers/AppealCrudController.php
@@ -5,9 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\CreateAppealRequest;
 use App\Models\Appeal;
 use App\Models\User;
-use Illuminate\Foundation\Application;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Routing\Redirector;
 
 class AppealCrudController extends Controller
 {
@@ -21,20 +19,50 @@ class AppealCrudController extends Controller
         return redirect('/appeals')->with('success', 'Обращение успешно создано');
     }
 
-    public function read(Appeal $appeal)
+    public function read(int $id)
     {
+        $appeal = Appeal::query()->find($id);
 
+        if (!$appeal) {
+            return redirect('/appeals')->with('error', 'Обращение не найдено');
+        }
+
+        if (!request()->user()->appeals->contains($appeal)) {
+            return redirect('/appeals')->with('error', 'У вас нет прав на просмотр этого обращения');
+        }
+
+        return view('pages.appeal.show')->with(compact('appeal'));
     }
 
-    public function update(CreateAppealRequest $request, Appeal $appeal): RedirectResponse
+    public function update(CreateAppealRequest $request, int $id): RedirectResponse
     {
+        $appeal = Appeal::query()->find($id);
+
+        if (!$appeal) {
+            return redirect('/appeals')->with('error', 'Обращение не найдено');
+        }
+
+        if (!request()->user()->appeals->contains($appeal)) {
+            return redirect('/appeals')->with('error', 'У вас нет прав на редактирование этого обращения');
+        }
+
         $appeal->update($request->validated());
 
         return redirect('/appeals')->with('success', 'Обращение успешно изменено');
     }
 
-    public function delete(Appeal $appeal)
+    public function delete(int $id): RedirectResponse
     {
+        $appeal = Appeal::query()->find($id);
+
+        if (!$appeal) {
+            return redirect('/appeals')->with('error', 'Обращение не найдено');
+        }
+
+        if (!request()->user()->appeals->contains($appeal)) {
+            return redirect('/appeals')->with('error', 'У вас нет прав на удаление этого обращения');
+        }
+
         $appeal->delete();
 
         return redirect('/appeals')->with('success', 'Обращение успешно удалено');
